fix: Rename types module to types_
